Migrating to the PIXEL version of Rasbian.
This will be based on PIXEL and the installed chromium. It will run
the cache in a ram disk. This ram disk will be created on boot. All uneeded
programs are removed.

This so far monitoring and the ram disk work.

Also firewall setup using UFW, and correct ports are allowed

Signed-off-by: Joseph Harry <findarato@gmail.com>